refactor(ui): move sidebar width variables to CSS variables.
