+++ packages/ui-react/src/components/sidebar/index.ts
@@ -1,5 +1,4 @@
 export * from './sidebar.css'
-export * from './sidebar-utils'
 export * from './sidebar-context'
 export * from './sidebar-main'
 export * from './sidebar-layout'
