+++ website/src/content/changelog/0.5.3.mdx
@@ -3,6 +3,8 @@ title: v0.5.3 (2025/03/27)
 ---
 
 - Changes: split [Sidebar](/docs/components/sidebar) component files.
+- Move sidebar width variables to CSS variables.
+- Move sidebar cookies parameters to SidebarProvider props.
 
 **Contributors:**
 [@riipandi](https://github.com/riipandi)
